+++ back_end/config/passport.js
@@ -1,11 +1,11 @@
-const JwtStrategy = require('passport-jwt').Strategy;
-const ExtractJwt = require('passport-jwt').ExtractJwt;
+const JwtStrategy = require('passport-jwt').Strategy
+const ExtractJwt = require('passport-jwt').ExtractJwt
 const User = require('../models/User').User
-const keys = require('./keys');
+const keys = require('./keys')
 
-const opts = {};
-opts.jwtFromRequest = ExtractJwt.fromAuthHeaderAsBearerToken();
-opts.secretOrKey = keys.secretOrKey;
+const opts = {}
+opts.jwtFromRequest = ExtractJwt.fromAuthHeaderAsBearerToken()
+opts.secretOrKey = keys.secretOrKey
 
 module.exports = passport => {
   passport.use(
@@ -13,11 +13,11 @@ module.exports = passport => {
       User.findById(jwt_payload.id)
         .then(user => {
           if (user) {
-            return done(null, user);
+            return done(null, user)
           }
-          return done(null, false);
+          return done(null, false)
         })
-        .catch(err => console.log(err));
+        .catch(err => console.log(err))
     })
-  );
-};
+  )
+}
