+++ back_end/models/JC.js
@@ -4,12 +4,12 @@ const Schema = mongoose.Schema
 const JCSchema = new Schema({
   jcFirstName: {
     type: String,
-    required: true
+    required: true,
   },
   jcLastName: {
     type: String,
-    required: true
-  }
+    required: true,
+  },
 })
 
 module.exports = JC = mongoose.model('JC', JCSchema)
