+++ back_end/models/User.js
@@ -1,35 +1,34 @@
-const mongoose = require('mongoose');
+const mongoose = require('mongoose')
 const Schema = mongoose.Schema
 
 var UserSchema = new Schema({
-    // _id: Schema.Types.ObjectId,
-    email: {
-        type: String,
-        required: true
-    },
-    username : {
-        type: String,
-        required: true
-    },
-    password: {
-        type: String,
-        required: true,
-    },
-    role: {
-        type: String,
-        required: true
-    },
-    activated : {
-        type: Boolean,
-        required : true
-    },
-    details: {
-        type: Object,
-        required: true,
-    },
-    hash: {
-        type: String
-    }
-});
+  email: {
+    type: String,
+    required: true,
+  },
+  username: {
+    type: String,
+    required: true,
+  },
+  password: {
+    type: String,
+    required: true,
+  },
+  role: {
+    type: String,
+    required: true,
+  },
+  activated: {
+    type: Boolean,
+    required: true,
+  },
+  details: {
+    type: Object,
+    required: true,
+  },
+  hash: {
+    type: String,
+  },
+})
 
-module.exports = User = mongoose.model('User', UserSchema);
+module.exports = User = mongoose.model('User', UserSchema)
